# api.yaml
openapi: 3.0.4
info:
  title: pyroaster API
  version: 3.0.0
servers:
  - url: http://192.168.1.90
    description: STA (router network)
  - url: http://192.168.4.1
    description: AP (direct connection)
tags:
  - name: Time
    description: Control the timer
  - name: Motors
    description: Control the motors
  - name: Controller
    description: Manage the controller
  - name: Configurations
    description: Saved roasting presets
  - name: Events
    description: Server-Sent Events stream
  - name: Reset
    description: Device reboot

paths:
  /time:
    post:
      tags: [Time]
      summary: Change the timer
      operationId: changeTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [add, reduce, change]
                time:
                  type: integer
                  minimum: 0
                  description: Required when action is "change". Time in seconds.
                  example: 60
      responses:
        '200':
          description: Updated timer state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Time'
        '400':
          description: Invalid action or missing/invalid time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /controller_config:
    get:
      tags: [Controller]
      summary: Get controller configuration
      operationId: getControllerConfig
      responses:
        '200':
          description: Current controller config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Controller'
    patch:
      tags: [Controller]
      summary: Update controller configuration
      operationId: changeControllerConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                starting_temperature:
                  type: integer
                  minimum: 0
                  example: 190
                time:
                  type: integer
                  minimum: 0
                  example: 9000
      responses:
        '200':
          description: Updated controller config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Controller'
        '400':
          description: Invalid field type or value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /controller:
    post:
      tags: [Controller]
      summary: Control the controller status
      operationId: handleControllerStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [activate, deactivate, stop]
      responses:
        '200':
          description: Updated controller config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Controller'
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /motors:
    post:
      tags: [Motors]
      summary: Control individual motors
      operationId: handleMotorStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                motor_a:
                  type: boolean
                  example: true
                motor_b:
                  type: boolean
                  example: false
                motor_c:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Current motor states
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Motor'
        '400':
          description: Invalid motor value (must be boolean)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /config:
    get:
      tags: [Configurations]
      summary: Get all saved presets
      operationId: getSavedConfigurations
      responses:
        '200':
          description: All saved presets as a dict keyed by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigMap'
        '500':
          description: Failed to read config file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Configurations]
      summary: Save a new preset
      operationId: createSavedConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, starting_temperature, time]
              properties:
                name:
                  type: string
                  example: cafe
                starting_temperature:
                  type: integer
                  minimum: 0
                  example: 190
                time:
                  type: integer
                  minimum: 0
                  example: 9000
      responses:
        '200':
          description: Updated config map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigMap'
        '400':
          description: Validation error or name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to read/write config file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /config/{name}:
    delete:
      tags: [Configurations]
      summary: Delete a saved preset
      operationId: deleteSavedConfiguration
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Name of the preset to delete
      responses:
        '200':
          description: Updated config map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigMap'
        '404':
          description: Config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to read/write config file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /events:
    get:
      tags: [Events]
      summary: SSE event stream
      description: >
        Streams four event types when data changes.
        Each event is a JSON object sent as an SSE message.
      operationId: startEventStreaming
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: object
                description: >
                  Events: "sensors" (Sensors), "time" (Time),
                  "states" (Motor), "controller" (Controller).
                  Only sent when data changes from the previous value.
              examples:
                sensors:
                  summary: sensors event
                  value:
                    temperature: 120
                    humidity: 80
                time:
                  summary: time event
                  value:
                    total_time: 9000
                    current_time: 4500
                states:
                  summary: states event
                  value:
                    motor_a: true
                    motor_b: false
                    motor_c: false
                controller:
                  summary: controller event
                  value:
                    starting_temperature: 190
                    time: 9000
                    status: "on"

  /reset:
    post:
      tags: [Reset]
      summary: Reboot the device
      operationId: resetMicro
      responses:
        '200':
          description: Reboot scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rebooting...

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "'field' is required"

    Sensors:
      type: object
      properties:
        temperature:
          type: integer
          description: Temperature in celsius
          example: 120
        humidity:
          type: integer
          description: Humidity percentage
          example: 80

    Time:
      type: object
      properties:
        total_time:
          type: integer
          example: 9000
        current_time:
          type: integer
          example: 4500

    Motor:
      type: object
      properties:
        motor_a:
          type: boolean
          example: true
        motor_b:
          type: boolean
          example: false
        motor_c:
          type: boolean
          example: false

    Controller:
      type: object
      properties:
        starting_temperature:
          type: integer
          minimum: 0
          example: 190
        time:
          type: integer
          minimum: 0
          example: 9000
        status:
          type: string
          enum: [on, off]
          readOnly: true
          example: off

    ConfigEntry:
      type: object
      properties:
        starting_temperature:
          type: integer
          minimum: 0
          example: 190
        time:
          type: integer
          minimum: 0
          example: 9000

    ConfigMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/ConfigEntry'
      example:
        cafe:
          starting_temperature: 190
          time: 9000
        mani:
          starting_temperature: 0
          time: 0